ci: add integration test to test targets
The integration test is now a Make target, 'integration_test'.
It is listed in the TEST_TARGETS variable and can be invoked
individually with `make integration_test`. All tests can be run
using `make test`. This works because `test` has TEST_TARGETS
as its dependencies.

To list all tests in your TEST_TARGETS:
	`make list_tests`